requirement.txt
asgiref==3.3.1
Django==3.1.5
django-appconf==1.0.4
django-imagefit==0.6.2
django-user-activity-log2==0.0.24
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
